Cleaned up a little and added param file
