+++ roscopter_sim/launch/mocap_landing.launch
@@ -53,7 +53,9 @@
 
 
   <group ns="m2u"> <!--m2u = motion capture to ublox conversion-->
-    <node name="mocap2ublox" pkg="roscopter" type="mocap2ublox2.py" output="screen">
+    <!-- Load common parameters -->
+    <rosparam command="load" file="$(find roscopter)/params/mocap2ublox.yaml"/>
+    <node name="mocap2ublox" pkg="roscopter" type="mocap2ublox_ros.py" output="screen">
       <remap from="gnss" to="/$(arg mav_name)/gps/data"/>
       <remap from="rover_mocap" to="/sim/rover_mocap"/>
       <remap from="base_mocap" to="/sim/base_mocap"/>
